Refactor neurotypes package into logical type groups
Split the monolithic interfaces.go into focused files:
- core_interfaces.go: Core architectural interfaces (Context, Service, Command, ServiceRegistry)
- command_types.go: Command system types (ParseMode, CommandArgs, HelpInfo)
- session_types.go: Session and conversation types (Message, SessionState, ChatSession)
- interfaces.go: Comprehensive package documentation and entry point

Benefits:
- Clear logical separation with single responsibility per file
- Enhanced maintainability and discoverability
- Comprehensive package documentation in interfaces.go
- Full backward compatibility maintained
- Consistent with existing model_types.go and theme_types.go pattern

All tests pass and existing imports continue to work unchanged.
